renamed tests to be more uniform, changed Response constructor to use Generic types instead of Object
